kill-frap: add figure/result for FACS of stable cell line.
This is a binary (pdf) used in the methods sectin of this chaper. We do have
the raw required to build this plot again, as we described in the README.md
file of the data directory. The problem is that we kinda lack the libraries
required to generate this plot via scripting.