Integrate H2AX review into the thesis appendices.
* thesis.tex: add the H2AX review bibtex file to our bibliography.
* appendices.tex: rename appendix with the title of the publication, and
add a short sentence on why it is here.
* H2AXreview.tex: a lot of changes to integrate on the thesis but without
changing content. Preamble of the paper was removed, including the author
names, title, and maketitle command. Replace the abstract command from
svmult with the abstract environment being used in the thesis. Replace the
\ugreekletter with \textgreekletter from the textgreek package being used
on the thesis. Remove the keyword and description which were used in the
publication to create a list of abreviations. Replace the \cite and
\citeasnoun command with \citep and \citet. Remove labels that were not
being used to reduce problems with the rest of the thesis and rename the
other labels with a 'h2ax-review'. Move figures and text from the bottom
of the paper to after their first reference. Replace \emph with our own
\species macro which specifically uses italic. Make use of siunitx package
macros like it is used in the rest of thesis. Also make use of \fref, \tref,
and \Sref instead of spelling out like "Fig.~\ref...". Use memoir's
sidcaption environment for the H2AX logo figure. Also make adjustments on
the table style, making use of \toprule, \midrule, etc instead of vertical
whitespace only using \svhline which was part of the svmult document class.