fancy-frap: register pre and post bleach frames to account for cell movement.
There is small movement in the time between the pre and post bleach
(1 minute) so difference between the two channels does not work.
Include a function that does the registration of the two channels.
This still causes problem because there are big blobs of stuff outside
the nucleus on the green channel that are very mobile and show up
on the difference between the channels.