fancy-frap: add foci coordinates from manually tracking.
I gave up on writing code to automatically track this foci because of the
bad signal to noise ratio, and bad temporal resolution (also complex because
the foci move out of the image sometimes, and some images have two foci).

So add the coordinates from manually tracking them as raw data.  Also add
the perl script we used to convert the MTrackJ tracking file to csv that
can be handled easily.