h2ax-review: set captions text for TOC.
Add shorted description to captions which is used for the table
of figures and tables.  By default the whole caption is used which
is too much.  Also cut down on the unecessary use of \protect (I
guess it was necessary when using the svmult document class).

Also, reformat lines that are too long.