histone-catalogue: add catalogue for mice as well as human.
Move the \input of files defining new commands out of the standard
catalogue preamble into its own "variables.tex" file.  This allows
"variables.tex" to be \input in the middle of the document (what we
want is to \input them into different groups).

The code used for the histone-catalogue creates variables named
the same for both organisms.  Solve this by including those sections
in group environment and only defining the commands within them.

figures/ and results/ within the histone-catalogue subtree now have
the organism name on the path so both human and mice analysis can
be done on the same tree.  So define new commands with the path on the
thesis and use those commands in the \input.

Move email option out of the histone-catalogue SConscript since that
is now parsed twice.