software: changes after Andrew corrections on Bioperl section.
* Pod::Weaver is about reducing POD, not documentation.  POD is
the markup language for the documentation, and we want to reduce
the ammount of that language, but not actually reduce the
documentation.
* We didn't package all Dist::Zilla plugins, only the ones needed.
* We don't download genes, we download sequences, including genomic
sequences.
* bp_genbank_ref_extractor is a script of Bio-EUtilities but is
not a module.  Saying that it's a component can be misleading.
If the point is saying from the start that it is in Bio-EUtilities
better just say that it was for them