software: changes after Andrew corrections on build-system section.
* "infrastructure environment" is never said.  Environment is enough
and what appears on the literature, but if we need to be more specific
then runtime is the word.
* It is not about making the environment available, it is about having
exact instructions to reproduce it (example, Debian as buildinfo files
which was developed for the reproducible-builds project).
* "software assemblies" is not used.  Builds is the word and makes it
obvious why it's called a build system.
* removed the "although it is very commonly used in Bioinformatics"
because SCons is a tool for software builds and does not need to cater
to the needs of bioinformatics data analysis.
* also, our tool does not enable the use of perl, only makes it slightly
easier with configuration options.